todo_backend/requirements.txt
# Lists dependencies like fastapi, uvicorn, and optionally pydantic.
fastapi
uvicorn[standard]
pytest
pytest-asyncio
httpx
sqlalchemy[asyncio]>=2.0
asyncpg
pydantic[email]
nats-py
